Split up sources
