+++ src/www/index.html
@@ -7,7 +7,7 @@
       font-family: system-ui;
     }
     body {
-      font-size: 2.0vmax;
+      font-size: 1.8vmax;
       color: white;
       background-color: #334;
     }
